moved stuff from index.js to app.js
